chore: 실행 가능한 JAR파일 생성 [#25]
* dependencies를 전부 포함하는 JAR도 생성
* Main class의 경로를 dev.min.skill에서 dev.mini.kakaoiopenbuilder.skill로 변경

Closed #25
