+++ src/main/java/dev/mini/kakaoiopenbuilder/skill/Main.java
@@ -1,0 +1,8 @@
+package dev.mini.kakaoiopenbuilder.skill;
+
+public class Main {
+    public static void main(String[] args) {
+        System.out.println("Hello Kakao I OpenBuilder Java Library!");
+    }
+
+}
