+++ src/main/java/dev/mini/kakaoiopenbuilder/skill/payload/intent/Intent.java
@@ -1,10 +1,11 @@
-package dev.mini.skill.payload.bot;
+package dev.mini.kakaoiopenbuilder.skill.payload.intent;
 
 import java.util.Objects;
 
-public class Bot {
+public class Intent {
     private String id;
     private String name;
+    private Extra extra;
 
     public String getId() {
         return id;
@@ -14,24 +15,29 @@ public class Bot {
         return name;
     }
 
+    public Extra getExtra() {
+        return extra;
+    }
+
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
-        Bot bot = (Bot) o;
-        return Objects.equals(id, bot.id) && Objects.equals(name, bot.name);
+        Intent intent = (Intent) o;
+        return Objects.equals(id, intent.id) && Objects.equals(name, intent.name) && Objects.equals(extra, intent.extra);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(id, name);
+        return Objects.hash(id, name, extra);
     }
 
     @Override
     public String toString() {
-        return "Bot{" +
+        return "Intent{" +
                 "id='" + id + '\'' +
                 ", name='" + name + '\'' +
+                ", extra=" + extra +
                 '}';
     }
