+++ src/main/java/dev/mini/kakaoiopenbuilder/skill/payload/intent/Reason.java
@@ -1,4 +1,4 @@
-package dev.mini.skill.payload.intent;
+package dev.mini.kakaoiopenbuilder.skill.payload.intent;
 
 import java.util.Objects;
