+++ src/main/java/dev/mini/kakaoiopenbuilder/skill/payload/userrequest/Block.java
@@ -1,11 +1,10 @@
-package dev.mini.skill.payload.intent;
+package dev.mini.kakaoiopenbuilder.skill.payload.userrequest;
 
 import java.util.Objects;
 
-public class Intent {
+public class Block {
     private String id;
     private String name;
-    private Extra extra;
 
     public String getId() {
         return id;
@@ -15,29 +14,24 @@ public class Intent {
         return name;
     }
 
-    public Extra getExtra() {
-        return extra;
-    }
-
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
-        Intent intent = (Intent) o;
-        return Objects.equals(id, intent.id) && Objects.equals(name, intent.name) && Objects.equals(extra, intent.extra);
+        Block block = (Block) o;
+        return Objects.equals(id, block.id) && Objects.equals(name, block.name);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(id, name, extra);
+        return Objects.hash(id, name);
     }
 
     @Override
     public String toString() {
-        return "Intent{" +
+        return "Block{" +
                 "id='" + id + '\'' +
                 ", name='" + name + '\'' +
-                ", extra=" + extra +
                 '}';
     }
