+++ src/main/java/dev/mini/kakaoiopenbuilder/skill/payload/userrequest/Properties.java
@@ -1,4 +1,4 @@
-package dev.mini.skill.payload.userrequest;
+package dev.mini.kakaoiopenbuilder.skill.payload.userrequest;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
