+++ src/main/java/dev/mini/kakaoiopenbuilder/skill/payload/userrequest/UserRequest.java
@@ -1,4 +1,4 @@
-package dev.mini.skill.payload.userrequest;
+package dev.mini.kakaoiopenbuilder.skill.payload.userrequest;
 
 import java.util.Objects;
