--- src/main/java/dev/mini/skill/Main.java
@@ -1,9 +1,0 @@
-package dev.mini.skill;
-
-public class Main {
-
-    public static void main(String[] args) {
-
-    }
-
-}
